+++ jest.config.ts
@@ -1,0 +1,10 @@
+import type { Config } from "@jest/types";
+
+const config: Config.InitialOptions = {
+  preset: "ts-jest",
+  rootDir: "test",
+  testEnvironment: "node",
+  verbose: true,
+};
+
+export default config;