+++ src/constant/query/GOOGLE_COLORS.ts
@@ -1,0 +1,16 @@
+const COLORS = [
+  "red",
+  "blue",
+  "purple",
+  "orange",
+  "yellow",
+  "green",
+  "teal",
+  "pink",
+  "white",
+  "gray",
+  "black",
+  "brown",
+];
+
+export default COLORS;